bug fixes
elements within the view are now positioned on full pixels (this problem
caused blurry ui elements in the message bubble), added missing image
files to the demo project

+++ BBMessageBubble/Source/RHMessageBubble.m
@@ -227,10 +227,10 @@
 	[activityIndicator release];
 	
 	CGRect spinnerFrame = activityIndicator.frame;
-	spinnerFrame.origin = CGPointMake(view.frame.size.width / 2 - 
-									  activityIndicator.frame.size.width / 2, 
-									  view.frame.size.height / 2 - 
-									  activityIndicator.frame.size.height / 2);
+	spinnerFrame.origin = CGPointMake((NSInteger)(view.frame.size.width / 2 - 
+									  activityIndicator.frame.size.width / 2), 
+									  (NSInteger)(view.frame.size.height / 2 - 
+									  activityIndicator.frame.size.height / 2));
 	activityIndicator.frame = spinnerFrame;
 	
 	[RHMessageBubble addBackground:view];
@@ -265,7 +265,7 @@
 		numLines = rounded;
 	}
 	
-	CGRect rect = CGRectMake(RH_MESSAGE_BUBBLE_PADDING, RH_MESSAGE_BUBBLE_PADDING, RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING, RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING);
+	CGRect rect = CGRectMake(RH_MESSAGE_BUBBLE_PADDING, RH_MESSAGE_BUBBLE_PADDING, (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING), (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING));
 	
 	UILabel* label = [RHMessageBubble getLabelWithText:text];
 	label.frame = rect;
@@ -303,7 +303,7 @@
 		numLines = rounded;
 	}
 	
-	CGRect rect = CGRectMake(0, 0, RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING, numLines * lineHeight.height);
+	CGRect rect = CGRectMake(0, 0, (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING), (NSInteger)(numLines * lineHeight.height));
 	
 	UILabel* label = [[UILabel alloc] initWithFrame:rect];
 	[label setText:text];
@@ -370,22 +370,22 @@
 
 + (void)addTopView:(UIView*)topView andBottomView:(UIView*)bottomView toView:(UIView*)view
 {
-	UIView* container = [[UIView alloc] initWithFrame:CGRectMake(0, 0, RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING, 1)];
+	UIView* container = [[UIView alloc] initWithFrame:CGRectMake(0, 0, (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING), 1)];
 	container.backgroundColor = [UIColor clearColor];
 	
 	NSInteger paddingBetween = 10;
 	
-	CGRect rect = CGRectMake(RH_MESSAGE_BUBBLE_PADDING, 0, RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING, topView.frame.size.height + bottomView.frame.size.height + paddingBetween);
+	CGRect rect = CGRectMake(RH_MESSAGE_BUBBLE_PADDING, 0, (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2 * RH_MESSAGE_BUBBLE_PADDING), topView.frame.size.height + bottomView.frame.size.height + paddingBetween);
 	container.frame = rect;
 	
-	NSInteger topViewCenter = (RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2*RH_MESSAGE_BUBBLE_PADDING) / 2 - topView.frame.size.width / 2;
+	NSInteger topViewCenter = ((NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2*RH_MESSAGE_BUBBLE_PADDING) / 2 - topView.frame.size.width / 2);
 	
 	topView.frame = CGRectMake(topViewCenter, 
 							   0, 
 							   topView.frame.size.width, 
 							   topView.frame.size.height);
 	
-	NSInteger bottomViewCenter = (RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2*RH_MESSAGE_BUBBLE_PADDING) / 2 - bottomView.frame.size.width / 2;
+	NSInteger bottomViewCenter = (NSInteger)((RH_MESSAGE_BUBBLE_SIDE_LENGTH - 2*RH_MESSAGE_BUBBLE_PADDING) / 2 - bottomView.frame.size.width / 2);
 	
 	bottomView.frame = CGRectMake(bottomViewCenter, 
 								  topView.frame.size.height + paddingBetween, 
@@ -397,10 +397,10 @@
 	
 	CGRect contRect = container.frame;
 	
-	contRect.origin.y = view.frame.size.height / 2 - container.frame.size.height / 2;
+	contRect.origin.y = (NSInteger)(view.frame.size.height / 2 - container.frame.size.height / 2);
 	
 	if (bottomView.frame.size.height + topView.frame.size.height > RH_MESSAGE_BUBBLE_SIDE_LENGTH - RH_MESSAGE_BUBBLE_PADDING * 2) {
-		contRect.size.height = RH_MESSAGE_BUBBLE_SIDE_LENGTH - RH_MESSAGE_BUBBLE_PADDING * 2;
+		contRect.size.height = (NSInteger)(RH_MESSAGE_BUBBLE_SIDE_LENGTH - RH_MESSAGE_BUBBLE_PADDING * 2);
 	}
 	
 	container.frame = contRect;
